requirements.txt
PyYAML
tweepy